Script to insert data
Recommendation: Create quotes from the MVC because if you insert quotes from the SMSS and then try to do so through the MVC, you will get an error because it will generate the quote code from 01. You will then have to try to create quotes until you exceed the number of quotes inserted directly into the database. This bug is pending resolution.